requirements.txt
torch
torchvision
transformers
datasets
numpy
pandas
matplotlib
jupyter
PySyft
pyyaml
pytest
syft
scipy
POT
prv-accountant
scipy
pydantic
opt_einsum
ml-swissknife
fairscale
tqdm
deepspeed
jupyter
jupyterlab
eth-brownie
web3